+++ .gitignore
@@ -23,7 +23,7 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
+yarn.lock
 
 ### Java ###
 # Compiled class file
