+++ .travis.yml
@@ -1,7 +1,19 @@
+
 language: node_js
-node_js:
-   - "9"
-   - "8"
-script:
-  - npm run lint
-  - npm run test
+
+before_install:
+- curl -o- -L https://yarnpkg.com/install.sh | bash -s -- --version 1.2.1
+- export PATH=$HOME/.yarn/bin:$PATH
+
+matrix:
+  include:
+  - node_js: '9.0'
+    script:
+      - yarn lint
+      - yarn test
+
+cache:
+  yarn: true
+  directories:
+    - node_modules
+    - .jest/cache
