+++ App.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { View, StatusBar } from 'react-native';
 import { Provider } from 'react-redux';
-import { configureStore } from './src/store/ConfigureStore';
-// import Router from './src/Router';
-import { NavBar } from './src/components/common';
-import CoinsList from './src/components/CoinsList';
-import colors from './src/config/colors';
+import { configureStore } from './src/App/Store/ConfigureStore';
+import { NavBar } from './src/App/Components/Common';
+import CoinsList from './src/App/Components/CoinsList';
+import colors from './src/App/Configs/Colors';
 
 const store = configureStore();
