+++ jsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+      "target": "ES6",
+      "module": "commonjs",
+      "allowSyntheticDefaultImports": true
+    },
+    "exclude": [
+      "node_modules",
+      ".vscode/typings"
+    ]
+  }
