+++ package.json
@@ -40,13 +40,13 @@
     "react-navigation": "^1.0.3",
     "react-redux": "^5.0.6",
     "react-sparklines": "^1.7.0",
-    "realm": "^2.2.7",
     "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
     "redux-saga": "^0.16.0",
     "redux-thunk": "^2.2.0",
     "reselect": "^3.0.1",
-    "styled-components": "^3.2.1"
+    "styled-components": "^3.2.1",
+    "uuid": "^3.2.1"
   },
   "devDependencies": {
     "babel-eslint": "^8.2.1",
@@ -56,6 +56,10 @@
     "react-native-scripts": "1.8.1",
     "react-test-renderer": "16.0.0",
     "redux-devtools-extension": "^2.13.2",
-    "rimraf": "^2.6.2"
+    "rimraf": "^2.6.2",
+    "tslint": "^5.9.1",
+    "typescript": "^2.7.2",
+    "typescript-styled-plugin": "^0.6.0",
+    "typescript-tools": "^0.3.1"
   }
 }
