+++ src/App/API/CryptocompareDocumentation.json
