+++ src/App/Actions/AllCoinsActions.js
