+++ src/App/Assets/logo.png
