+++ src/App/Assets/logo@2x.png
