+++ src/App/Assets/logo@3x.png
