+++ src/App/Assets/settings-icon@3x.png
