+++ src/App/Components/CoinItem/CoinName.js
