+++ src/App/Components/CoinItem/CoinPrice.js
