+++ src/App/Components/CoinItem/SparkLines.js
