+++ src/App/Components/CoinsList.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import { View, FlatList } from 'react-native';
+import { FlatList } from 'react-native';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { getBriefCoinData } from '../actions';
+import { getBriefCoinData } from '../Actions';
 import CoinItem from './CoinItem';
 
 class CoinsList extends Component {
