+++ src/App/Components/Common/AddCoinButton.js
