+++ src/App/Configs/Colors.ts
@@ -1,0 +1,6 @@
+export enum Colors { 
+    greenblue = '#21ce99',
+    dark = '#15191e',
+    cellSeparator = '#4c4c4c',
+    textHeader = '#ffffff'
+}
