+++ src/App/Configs/Devices.ts
@@ -1,0 +1,6 @@
+export enum DeviceSizes {
+  small = 576,
+  medium = 768,
+  large = 992,
+  extra = 1200
+}
