+++ src/App/Configs/Fonts.ts
@@ -1,0 +1,3 @@
+export enum Fonts {
+    header = 'SFUIDisplay'
+}
