+++ src/App/Reducers/DefaultCoinsReducer.js
@@ -1,8 +1,8 @@
 import {
-  ALL_COINS_DATA,
-  ALL_COINS_DATA_FETCH_SUCCESS,
-  ALL_COINS_DATA_FETCH_FAIL
-} from '../actions/types';
+  DEFAULT_COINS_DATA,
+  DEFAULT_COINS_DATA_FETCH_SUCCESS,
+  DEFAULT_COINS_DATA_FETCH_FAIL
+} from '../Actions/types';
 
 const INITIAL_STATE = {
   data: null,
@@ -12,11 +12,11 @@ const INITIAL_STATE = {
 
 export default (state = INITIAL_STATE, action) => {
   switch(action.type) {
-    case ALL_COINS_DATA:
+    case DEFAULT_COINS_DATA:
       return { ...state, loading: true, error: '' };
-    case ALL_COINS_DATA_FETCH_SUCCESS:
+    case DEFAULT_COINS_DATA_FETCH_SUCCESS:
       return { ...state, ...INITIAL_STATE, data: action.payload };
-    case ALL_COINS_DATA_FETCH_FAIL:
+    case DEFAULT_COINS_DATA_FETCH_FAIL:
       return { ...state, error: 'Can not fetch data!', loading: false };
     default:
       return state;
