+++ src/App/Reducers/SearchResultsReducer.js
