+++ src/App/Reducers/SelectedCoinReducer.js
