--- src/actions/types.js
@@ -1,17 +1,0 @@
-export const ALL_COINS_DATA = 'all_coins_data';
-export const ALL_COINS_DATA_FETCH_SUCCESS = 'all_data_coins_fetch_success';
-export const ALL_COINS_DATA_FETCH_FAIL = 'all_data_coins_fetch_fail';
-
-export const BRIEF_COIN_DATA = 'brief_coin_data';
-export const BRIEF_COIN_DATA_FETCH_SUCCESS = 'brief_coin_data_fetch_success';
-export const BRIEF_COIN_DATA_FETCH_FAIL =  'brief_coin_data_fetch_fail';
-
-export const DEFAULT_COINS_DATA= 'default_coins_data';
-export const DEFAULT_DATA_FETCH_SUCCESS = 'default_fetch_success';
-export const DEFAULT_DATA_FETCH_FAIL = 'default_coins_fetch_fail';
-
-export const COIN_SPARKLINES_FETCH_SUCCESS = 'coin_sparklines_fetched_success';
-export const COIN_SPARKLINES_FETCH_FAIL = 'coin_sparklines_fetched_fail';
-
-export const COIN_CHART_DATA_FETCH_SUCCESS = 'coin_chart_data_fetched_success';
-export const COIN_CHART_DATA_FETCH_FAIL = 'coin_chart_data_fetched_fail';
