--- src/config/devices.js
@@ -1,6 +1,0 @@
-export default {
-  small: 576,
-  medium: 768,
-  large: 992,
-  extra: 1200,
-};
