+++ tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+      "allowJs": true,
+      "allowSyntheticDefaultImports": true,
+      "baseUrl": "src",
+      "experimentalDecorators": true,
+      "jsx": "react",
+      "module": "es2015",
+      "moduleResolution": "node",
+      "noUnusedLocals": true,
+      "noUnusedParameters": true,
+      "outDir": "build",
+      "plugins": [{ "name": "typescript-styled-plugin" }],
+      "pretty": true,
+      "strictNullChecks": false,
+      "target": "es6",
+      "types": ["jest", "react", "react-native"],
+      "noEmit": true
+    },
+    "include": ["typings/*.d.ts", "src/**/*.ts", "src/**/*.tsx"],
+    "exclude": [
+      "node_modules"
+    ]
+  }
