+++ tslint.json
@@ -1,0 +1,15 @@
+{
+    "extends": ["tslint:recommended", "react", "tslint-config-prettier"],
+    "rules": {
+      "arrow-parens": false,
+      "interface-name": [true, "never-prefix"],
+      "max-classes-per-file": [false],
+      "member-access": [false, "check-accessor", "check-constructor"],
+      "no-console": [true, ["error", ["warn", "error"]]],
+      "no-unused-variable": [true, { "ignore-pattern": "^_" }],
+      "object-literal-sort-keys": false,
+      "ordered-imports": true,
+      "switch-default": false,
+      "variable-name": [true, "allow-pascal-case", "allow-leading-underscore"]
+    }
+  }
